Extracted checking arguments into separate method
Improvements also on checking if composer is executable
- ability to work also on Windows.

Now all checks are after parsing arguments so also default
values of the script are checked.

In case when default "composer" binary doesn't exist script
shouldn't be executed. The same in case when default
modulesPath is not a correct directory, script stops with
error message.